feat: add an optin for the iframe.html to be loaded with instrument=true
In order for storybook's instrumenter to be run on the iframe.html page,
the GET parameter instrument=true must be included on the page. The
test-runner addon currently does not provide this option when running the
tests and therfore the instrumenter does not run under test-runner.

This change adds the option of enabling the instrumenter, using the
`process.env.INSTRUMENT` option, which adds the `instrument=true` GET
variable to the iframe URL.