fix: gracefully kill process on SIGINT and exit
SIGINT is a signal that is generally used to indicate a program
is being interrupted by the user (like when entering "ctrl + c") and
allows a program to be gracefully terminated.

The event handler for `SIGINT` (as well as exit) does not actually
kill the process so when the user tries to interrupt the program,
the process is left hanging which is not what is expected for CLI
programs in general. As it is right now, the entire test suite needs
to complete running before the program exits.

This commit is simply killing the process upon a SIGINT or exit event.